refactor(core/geometry): split implementation files
